Merging changes up to revision: 12063 in branch svn://ocelot/blur/branches/blurdev/blurdev-linux-2011-07-19 Preference now checks if the xml was loaded successfully this way if a preference file exists, but is corrupt it will populate with defaults. LoggerWindow has a menu item to set the indentations use tabs setting of the DocumentEditor.
runDelayed now will process as many items in the queue that it can within a set amount of time, default value of 0.1seconds.
runDelayed will catch the RuntimeError 'underlying C/C++ object has been deleted'.
If any other exception happens it should still raise the exception, but not break the queue.(items in the queue, but postQueueEvent was not called.)
Bug: 6855


git-svn-id: svn://svn.blur.com/blur/trunk/python/blurdev@12064 896c8824-d3f5-0310-9c61-ac502b9a79dc