Process pth files instead of egg-link files
Pth files are how egg-link installs actually get added to python, the
egg-link file is just for pip, python uses the .pth file.

Re-implement site.addsitedir so we can handle pth files defined on a
different operating system. This allows us to setup a treegrunt
virtualenv with egg-links on linux but still use it on windows.