Pinning blurdev requirements to their latest, working version.
Using third-party packages we run the risk of them releasing a version
that proves in-compatible with our usage and/or environment. For example,
in our `blur-dists`-package, we require `pytest`. They recently released
an update that deprecated Python 2 in favor of Python 3. This resulted
in failed installations of our package software distributions. By pinning
releases to confirmed compatible versions, we can audit future updates to
our dependencies (i.e. security patches, new features) and manually
opt-in to them as we deem fit.