Fix getting errorMessageColor in ErrorDialog from previous commit
The value of a pyqtProperty(PySideX untested) are only accessible
from instances, not class properties. Even if this wasn't the case, it
likely would only get set to the style sheet setting when initialized as well.

The ErrorDialog now has its own Property that is styled the same by default.