Show a warning if unable to run instance_shutdown
This can happen if the Qt application was closed or killed before instance_shutdown
is called. It will now show a warning instead of raising an error to make
it easier to debug why your prefs may not be saving.

This helps with our Nuke integration which we haven't found a good way to
detect the app shutting down before our widget is deleted.