Better handling of writes to __stderr__/__stdout__
On windows, if using pythonw.exe these streams are set to None. If you try
to print to None it will write to `sys.stdout` which causes recursion.
The intent of these calls is to write to the host shell when errors happen
processing a normal PrEditor write so the problem can be debugged.

If you run into that issue when using pythonw.exe you will need to switch
to python.exe to see the debug output.